+++ resources/hpxml-measures/workflow/tests/util.rb
@@ -375,6 +375,10 @@ def _verify_outputs(rundir, hpxml_path, results, hpxml, unit_multiplier)
     if timestep > 15
       next if message.include?('Timestep: Requested number') && message.include?('is less than the suggested minimum')
     end
+    # Location doesn't match EPW station
+    if hpxml_path.include? 'base-location-detailed.xml'
+      next if message.include? 'Weather file location will be used rather than entered (IDF) Location object.'
+    end
     # TODO: Check why this house produces this warning
     if hpxml_path.include? 'house044.xml'
       next if message.include? 'FixViewFactors: View factors not complete. Check for bad surface descriptions or unenclosed zone'